Refactor models into db/models
