+++ src/db/models/Template/index.js
@@ -1,5 +1,5 @@
 import { z } from 'zod'
-import { db } from "../init.js"
+import { db } from "/src/db/init.js"
 import {
   doc,
   getDocs,
