+++ src/db/models/Template/utils.js
@@ -1,6 +1,6 @@
 import cronParser from 'cron-parser'
 import { DateTime } from 'luxon'
-import { db } from "../init.js"
+import { db } from "/src/db/init.js"
 import { 
   updateDoc, 
   doc, 
@@ -12,7 +12,7 @@ import {
   deleteDoc, 
   writeBatch 
 } from 'firebase/firestore'
-import { getRandomID } from '../../utils/core.js'
+import { getRandomID } from '/src/utils/core.js'
 
 const { parseExpression } = cronParser
