+++ src/lib/Reusable/FlexibleDayTask.svelte
@@ -46,7 +46,7 @@
   // Assumes `task` is hydrated
   import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import Checkbox from './Checkbox.svelte'
-  import Task from '/src/db/Task.js'
+  import Task from '/src/db/models/Task.js'
 
   export let task = null
   export let pixelsPerHour = null
