+++ src/lib/Reusable/IconTaskElement.svelte
@@ -101,7 +101,7 @@
   import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import DoodleIcon from '$lib/Reusable/DoodleIcon.svelte'
   import { pixelsPerHour } from '/src/lib/Calendar/store.js'
-  import Task from '/src/db/Task.js'
+  import Task from '/src/db/models/Task.js'
 
   export let task = null
