+++ src/lib/TaskPopup/PhotoUpload.svelte
@@ -24,7 +24,7 @@
   import { getRandomID, getTimeInHHMM } from '/src/utils/core.js'
   import { DateTime } from 'luxon'
   import { deleteImage } from '/src/db/helpers.js'
-  import Task from '/src/db/Task.js'
+  import Task from '/src/db/models/Task.js'
 
   export let taskObject
