+++ src/lib/TaskTree/Dropzone.svelte
@@ -18,7 +18,7 @@
   import { user, activeDragItem } from '/src/store'
   import { increment, writeBatch, doc } from 'firebase/firestore'
   import { db } from '../../db/init'
-  import Task from '/src/db/Task.js'
+  import Task from '/src/db/models/Task.js'
   import { HEIGHTS } from '/src/utils/constants.js'
   import { listTreeNodes } from '/src/db/treeISOs.js'
