+++ src/lib/Templates/Templates.svelte
@@ -3,7 +3,7 @@
   import { onMount } from 'svelte'
   import { user, calendarTasks } from '/src/store'
   import { templates } from './store.js'
-  import Template from '/src/db/Template'
+  import Template from '/src/db/models/Template'
   import { filterByType } from './utils.js'
 
   $: weeklyTasks = filterByType($templates, 'weekly')
